refactor(application): modified project folder structure with multiple file structure
