+++ app/main.py
@@ -1,10 +1,8 @@
 from fastapi import FastAPI
 from .router import task
+from app.router import base, task
 
 app = FastAPI()
 
-app.include_router(task)
-
-@app.get("/")
-def home_view():
-    return {"Hello": "World"}
+app.include_router(base.router)
+app.include_router(task.router)
