+++ app/router/base.py
@@ -1,0 +1,10 @@
+from fastapi import APIRouter
+
+router = APIRouter(
+    tags=["base"]
+)
+
+
+@router.get("/")
+def root():
+    return {"Hello": "World"}
